refactor(accounting): extract TransactionalReportService from AccountingService (Wave-4 P5)
Behavior-preserving: getProfitLossReport, getMonthlyPLSummary, getComparativePL,
getBalanceSheet, getCashFlowStatement + private aggregateFinanceExpenses moved
verbatim into a new TransactionalReportService; AccountingService delegates
(signatures + controller + ReportsService wiring unchanged). Read-only, no
shared helpers. Existing tests pass via the facade; getComparativePL spy +
aggregateFinanceExpenses tests re-pointed to the new service instance (assertions
unchanged). reports gating spec (includeFinanceExpenses) still green.
