+++ apps/api/src/modules/accounting/accounting.module.ts
@@ -4,6 +4,7 @@ import { AccountingClosingController } from './closing.controller';
 import { AccountingService } from './accounting.service';
 import { PeakExportService } from './peak-export.service';
 import { ReceivablesReportService } from './receivables-report.service';
+import { TransactionalReportService } from './transactional-report.service';
 import { BadDebtService } from './bad-debt.service';
 import { BadDebtProvisionCron } from './bad-debt-provision.cron';
 import { MonthlyCloseService } from './monthly-close.service';
@@ -22,6 +23,7 @@ import { PeakModule } from '../peak/peak.module';
     AccountingService,
     PeakExportService,
     ReceivablesReportService,
+    TransactionalReportService,
     BadDebtService,
     BadDebtProvisionCron,
     MonthlyCloseService,
